long long not supported in c90 use long baby

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,16 +8,16 @@
 
 int main(void)
 
-{	
+{
 	int i;
-	long long f = 1, s = 2, n;
+	long f = 1, s = 2, n;
 
-	printf("%lld, %lld, ", f, s);
+	printf("%ld, %ld, ", f, s);
 
 	for (i = 3; i <= 50; i++)
 	{
 	n = f + s;
-	printf("%lld, ", n);
+	printf("%ld, ", n);
 	f = s;
 	s = n;
 	}
